서버 timestamp, tick 추정 함수들 자료형 수정, framedata 인자 변경 방지

+++ NGP_TermProject/NGPTermProject/NetworkSyncManager.h
@@ -12,9 +12,9 @@ class NetworkSyncManager
 {
 public:
 	uint64_t GetTimestampMs();
-	uint64_t GetEstimatedServerTimeMs();
-	uint64_t GetDelayedServerTick();
-	uint64_t GetEstimatedServerTick();
+	float GetEstimatedServerTimeMs();
+	float GetDelayedServerTick();
+	float GetEstimatedServerTick();
 	inline uint64_t GetUpdatedTick() { return updatedTick; }
 	bool UpdateServerTick();
 	void UpdateSyncData(const uint64_t clientSendTimestamp, const uint64_t serverSendTimestamp);
